+++ src/zoid/venmo/component.jsx
@@ -27,7 +27,7 @@ import {
   inlineMemoize,
 } from "@krakenjs/belter/src";
 import { FUNDING } from "@paypal/sdk-constants/src";
-import { SpinnerPage, Overlay } from "@paypal/common-components/src";
+import { SpinnerPage, VenmoOverlay } from "@paypal/common-components/src";
 
 import { getSessionID } from "../../lib";
 import { containerContent } from "../content";
@@ -74,13 +74,10 @@ export function getVenmoCheckoutComponent(): VenmoCheckoutComponent {
         prerenderFrame,
         props,
       }) => {
-        const {
-          nonce,
-          locale: { lang },
-        } = props;
-        const content = containerContent("Venmo")[lang];
+        const { nonce } = props;
+        const content = containerContent("venmo").en;
         return (
-          <Overlay
+          <VenmoOverlay
             context={context}
             close={close}
             focus={focus}